Add support for controlling metadata rendering in the query (#278)
* refactor query & parsing

Refactor query parsing to be more reusable and structured. Also convert
the sorting to an enum to make it easier to use.

* add 'show' to query definition and parser

This enum will be define which elements of the task to render in a given
query. By default, we render all of them (due date, tags, project,
description).

* make query available via Svelte context

This means we can avoid prop-drilling it

* use query show parameters in task renderer

This allows finer grained control of what gets rendered in the task list

* misc bugfix around datetime parsing

If there was no datetime, then some of the parsing would fail

* update readme & changelog