chore: make task.project and task.section more robust
All tasks will have a project, rather than allowing each
user to figure out how to handle undefined, let's preallocate
an unknown project for it if we fail the lookup.

Similarly, with sections, tasks may not have a section which
meant we could not distinguish between a failed lookup and
no section at all. Now we can.