chore: suppress on-click exit events on mobile
We were having issues with the semantics of modals as implemented in Obsidian.

In particular, the semantics where clicking outside the modal closes it. It seems
there might be some double-event shenanigans on mobile where clicking on the dialog
closes the dialog and then triggers the "out of modal" click handler. So I've
suppressed this behaviour by removing the element the on-click is on and re-adding
it. Its very dumb, but it works!