chore: fix biome configuration on NixOS
This commit fixes two problems. The first is that the default linux binary doesn't work
on NixOS. So we set BIOME_BINARY explicitly to the musl variant if we are on a x64 linux
system.

The other side of the coin is the biomejs-vscode extension. This has an argument for the binary
which is great! But the loading of the environment and the initialisation of the biomejs
extension is racey, so its unreliable. To fix this, I've added a 'patch-vscode-settings' script
to the shell which will fixup the workspace settings so this will work