plugin: fix duplicate task content rendering
If the task content was edited and then the query component refreshed,
we would see both the previous and edited versions of the task content.
This was because we would not clear the previous task content.

There are a number of ways to resolve this, but the simplest is to key
the component so we force a full remount when the content changes.